Properly set lockd port on CentOS 7
The upstream cookbook doesn't seem to properly set the lockd port to what we
want for CentOS 7. It looks as though upstream changed it so that you need to
use either/both setting the kernel module and the sysctl setting. For now, let's
set both. We can't reload the module if it's already in use but if we can change
the port of the running system with sysctl as long as we restart the nfs server
after changing that.

I also improved the ServerSpec tests to check to make sure the ports for the
various services are running on the ports we expect. Otherwise we run into
problems with clients.